18AL: Add support for historical objective (#961)
Introduce an ability Destination that can be used to specify
a destination hex. This ability can be used in e.g. 1822,
1844, 1856, 18AL.

Add two new accessors to the ability base class:
 - name, can be set during initialization, and changed later
 - value, set to nil during initialization, and changed later

When displaying corporation charter a section for all
abilities a company has that is not part of any company.
The name and value is displayed for all these abilities.

Configured historical objective hex for all corporations in 18AL
using the destination ability. When a corporation places its
token in its destination hex, it receives $100 to treasury as
it has reached its historical objective.

For historical objectives the name is "Historical objective"
and the value is set to the city name of the destination.
The value does not change upon reaching the destination;
that would be a possible improvement.

----

Also added handling of tile lay in Montgomery to ensure that
the tile will have the label 'M' to restrict the green and
brown tiles to M tiles. From the start Montgomery does not
have any label to allow for any yellow city tiles to be laid
there.

This change was part of an earlier commit but seems to have
been lost during migration.