18AL: Implement company Brown & Sons Lumber Co
Add the posibility to configure skip of feature check when
laying tile.

Add ability for Brown & Sons Lumber Co to lay tile 445
as a one of ability, for free.

Tile 445 does now ignore town count in the hex it is placed.
This tile will only be placed via the Brown & Sons Lumber
Co special ability.

A similar ability exists in 18MEX, but that will be implemented
later.