18AL/18MS: Implement company that gives discount for some trains
Configurable attribute 'train_discount' added. When this is set
for a corporation, the trains that has been configured in the
attribute, will appear as extra lines in the BuyTrains view,
shown with the discounted price.

When train bought for a discounted price the company closes
(if count is zero)

This attribute has been configured for:
- 18AL: New Decatur Yards (50% discount for 3, 4, 5 trains)
- 18MS: Mobile & Ohio Railway (100$ discount for 3+, 4+ trains)