18AL: Implement M&C RR and S&NA companies
Added two abilities that could be made generic but use
specific implementations for 18AL:
1. route_bonus, intended to give bonus once if all
   hexes in a specified route are visit in any route
2. hex_bonus, intended to give bonus each time a hex
   is visited in any route

When company Memphis & Charleston RR is bought a route_bonus
ability is now added to the purchasing corporation.
This ability is lost only if purchasing corporation changes
president. The route_bonus ability correspond to the Robert
E Lee and Pan American chits in 18AL but does not have to
be assigned to any trains, but instead assumed to be assigned
to any train that accomplises one of the routes.
The ability appear on the corporation charter, showing the
assigned hex.

Configured assign_hex ability for company South & North Alabama
RR. When one of the allowed hexes is assigned an hex it will add
a hex_bonus ability to the purchasing corporation. This ability
gives a $10 revenue bonus to any route that visits assigned hex.
The ability is removed when 3 trains rusts (ie phase 6).
The ability appear on the corporation charter, just showing
the $20 / $40 bonus information.

Log texts appear when route bonus is added to revenue. It is
not shown for any train route. Hex bonus is added to route
calculation and is not shown in any other way.