Added GAME_END_CHECK which is used to configure how to handle different game end triggers. Also redesigned the game end check part of the code to make it more general.
Added handling of stock market game end trigger.
Implement handling of game end during current OR.

Configure 18AL and 18MEX to use stock market end game trigger, and
to end in current OR if bank runs out or stock market maxed.

Corrected the displayed par prices when starting a corporation.
Now they are show in ascending order.

Make it possible to configure stock market prices as end game
triggers (by adding "e" in the configuration of the price), and
color these with a suitable light blue background. The blue
was selected to allow for adding a lighter blueish for 1849
Phase 16 boxes. As this color is not used for any alpha or
later games at the moment, the 18AL will be a good candidate
to test the choice of color.