Improve game end: current OR and/or stock market reach max
Added GAME_END which can be set to:
 :full_round    - Complete full OR set when end game triggered
 :current_round - Complete current OR where end game triggered
Default value is :full_round

Added GAME_END_CHECK_STOCKMARKET which is a boolean, default false.

Implemented a check for stock market max price, and use this
for checking end game condition (if GAME_END_CHECK_STOCKMARKET
is true).

Change so that if GAME_CHECK is :current_round a game end
triggered game will end in the current OR.

Configure 18AL to use stock market end game trigger, and
to end in current OR if bank runs out or stock market maxed.

Improved IPO price order so that it is always ascending.

TODO: There are more Mark Derrick games that could use this
configuration.
TODO: No test has been implemented for this.