18MEX: Correct configuration and add phase related events.
Correct quite a lot of configuration for trains and phases.

Split 3T (6) to 3T (4) and 3½T (2).

Split 6T (2) to 6T (1) and 6½T (1).

Change 4T to rust on 6½T purchase.

Add phase 3½, which is a copy of phase 3, but with a close
minors event.

Add an NdM merger event for phase 5.

Add phase 6½, which is a copy of phase 6.

Add event description for events specific to 18MEX.
Add trivial event handling methods - they just write out
something and have no other effect as of now.