Make it possible to change always_market_price
This is needed for titles what change base capitalization with
phase changes, e.g. 1856 and 18SJ.

In 18SJ corporations start as incremental, so always_market_price
is true. But if pared after phase 5 they are instead full cap.
So the corporations capitalization is then set to :full and
always_market_price is changed to false.