[18Rhl] Give error for Route to/from out-tokened off-board hex
Previously the player was allowed to do the route. But this means
that a player might not notice that a previously OK route now
pays 0 for the off-board hex. It could be argued that you want to
do the route anyway (to collect towns in between) but in almost
all cases you would prefer to re-route. So the choice was to go
for the normal case, and not cater for the rare special case.